chatbot/requirements.txt
numpy 
streamlit
python-dotenv
google-generativeai
pillow 